+++ csharp/Svix/Models/SvixClientOptions.cs
@@ -1,0 +1,11 @@
+﻿namespace Svix.Models
+{
+    public sealed class SvixClientOptions
+    {
+        public string AccessToken { get; set; }
+        
+        public string ServiceUrl { get; set; }
+
+        public bool Throw { get; set; }
+    }
+}
